Added basic rendering of seed points
This commit adds basic framework for the MapGen application. It creates
a basic main.cpp file that creates an instance of the MapGen app and runs
it. The MapGen app has the initialization, basic rendering framework,
loop, and destruction of the app. Added a map class that is used to
generate the map specifically. As of now the class only renders basic
seed points to use as a basis for Voronoi cells in the future. Also
implemented the start of a MathUtils class that will hold important
basic structs to be used in the app lie Vector2's for example.