Delaunay triangle implementation.
Expanded upon the MathUtils file to include structs for dealing with
Delaunay triangulation. It is the first step in the process of setting
up Voronoi cells and needed to be manually written to work with Vector2's
as a structure. Edges and Triangles were added to help with this. Also
added some seperation of the rendering in the Map class to draw both
the generated grid seed points and the Delaunay triangles as wanted.

Also updated the README.md to reflect some changes that were made and
lay out some install instructions for users.